+++ PokemonChoice.hpp
@@ -1,0 +1,11 @@
+#ifndef POKEMONCHOICE_HPP
+#define POKEMONCHOICE_HPP
+
+enum class PokemonChoice {
+    Charmander = 1,
+    Bulbasaur,
+    Squirtle,
+    Pikachu
+};
+
+#endif
