+++ Utility.hpp
@@ -1,6 +1,10 @@
+#ifndef UTILITY_HPP
+#define UTILITY_HPP
 class Utility {
 public:
     static void clearConsole();
     static void waitForEnter();
     static void clearInputBuffer(); // New helper function
 };
+
+#endif
